moved ServerStore and added utility function
